REF: separate Open Graph image selection from the plot directive
og:image selection had no real reason to live in plot_directive.py or to care
what produced an image. Moved it into its own pyvista/ext/_opengraph_image.py,
matching the og:description module: no imports from plot_directive, no
filtering by which extension rendered an image. It now numbers every image
node on the page -- a pyvista-plot, a plain `.. image::`, anything -- and lets
a page point at the Nth one. plot_directive.py's setup() still turns it on by
default, which is the only thing that couples them.

Renamed alongside the move, since neither name should imply the plot
directive:
* pyvista_plot_opengraph -> pyvista_opengraph_image, matching the sibling
  pyvista_opengraph_description.
* .. pyvista-plot-thumbnail:: -> .. pyvista-opengraph-thumbnail::

Its "renders fewer images than expected, don't warn" case still checks
PyVista's own pyvista_plot_skip / pyvista_plot_skip_optional, but only
through getattr with a default, so the module still works standalone without
plot_directive loaded -- it just can't explain that particular mismatch as
well.

Moved the Open Graph documentation out of plot_directive.rst into its own
doc/source/extras/opengraph.rst for the same reason, cross-linked from both
directions.

Verified image selection off a bare `.. image::` with no plot directive
involved at all, and confirmed the `:ref:` between the two doc pages resolves.
