+++ doc/source/extras/opengraph.rst
@@ -1,0 +1,110 @@
+.. _opengraph_docs:
+
+Open Graph link previews
+=========================
+
+When a documentation link is shared, the preview that appears alongside it comes
+from the page's `Open Graph <https://ogp.me>`_ metadata. PyVista fills that in
+automatically for every page: the image is an image the page shows, and the
+description is the page's own leading prose.
+
+This needs no configuration beyond enabling
+`sphinxext-opengraph <https://github.com/wpilibsuite/sphinxext-opengraph>`_ and
+telling it where the documentation is published:
+
+.. code-block:: python
+
+    extensions = [
+        "pyvista.ext.plot_directive",
+        "sphinxext.opengraph",
+    ]
+
+    ogp_site_url = "https://docs.example.org/"
+
+Both features follow ``sphinxext.opengraph``: ``pyvista_opengraph_image`` and
+``pyvista_opengraph_description`` default to ``None``, which means "on if
+``sphinxext.opengraph`` is enabled". Set either to ``True`` to require it -- the
+build then fails if ``sphinxext.opengraph`` is missing -- or to ``False`` to opt
+out. Enabling :mod:`pyvista's plot directive <pyvista.ext.plot_directive>` turns
+both on by default, but neither depends on it: a page previews whichever image it
+shows, from any source, and its description is built the same way whether or not
+the page uses the plot directive at all.
+
+Choosing the preview image
+---------------------------
+
+By default a page previews the first image it shows. Pages where the first image
+is only scene-setting can select another one with the
+``pyvista-opengraph-thumbnail`` directive:
+
+.. code-block:: rst
+
+   .. pyvista-opengraph-thumbnail:: 2
+
+The argument is the one-based position of the image among *all* images on the
+page, in the order they appear -- it does not matter what produced them, so this
+works the same next to a ``pyvista-plot``, a plain ``.. image::``, or anything
+else. It counts images, not files, so it is unaffected by how the generated
+filenames happen to be numbered. Negative values count backwards from the last
+image.
+
+The directive renders nothing and can go anywhere on the page, which means it can
+sit next to the code it refers to instead of at the top. In a docstring the natural
+place is the start of the ``Examples`` section:
+
+.. code-block:: rst
+
+   Examples
+   --------
+   .. pyvista-opengraph-thumbnail:: 2
+
+   Create a sphere.
+
+   >>> import pyvista as pv
+   >>> pv.Sphere().plot()
+
+   Clip it, which is what this page is really about.
+
+   >>> pv.Sphere().clip().plot()
+
+A page has a single ``<head>``, so it has a single link preview -- section anchors
+cannot have their own. Using the directive twice on one page therefore warns and
+keeps the first selection. This can happen without either docstring being wrong, on
+pages that document several objects at once with ``:members:``.
+
+Selecting an image the page does not have also warns, and falls back to the first
+image. That is silent while PyVista's own ``pyvista_plot_skip`` or
+``pyvista_plot_skip_optional`` is enabled, since those builds deliberately render
+fewer images.
+
+Sphinx-Gallery examples
+------------------------
+
+Gallery examples already have a thumbnail, and their preview always matches it, so
+that a shared link shows the same picture as the gallery. PyVista uses the full
+resolution version of that image rather than the gallery's own thumbnail file, which
+is too small to preview well.
+
+Using ``pyvista-opengraph-thumbnail`` in a gallery example is an error. Use
+Sphinx-Gallery's own selector instead:
+
+.. code-block:: python
+
+   # sphinx_gallery_thumbnail_number = 2
+
+Preview descriptions
+----------------------
+
+``sphinxext-opengraph`` builds its description from every piece of text on a page
+until it has enough characters, which works poorly for generated pages. Autodoc
+output is invisible to it -- Sphinx wraps docstrings in a node it treats as an
+admonition and skips -- so API pages end up described by whatever sits outside that
+node, or not described at all. Sphinx-Gallery pages pick up download links, the
+timing footer and the "Gallery generated by Sphinx-Gallery" signature.
+
+PyVista instead uses the page's leading paragraphs of actual prose, up to
+``ogp_description_length`` characters, skipping signatures, parameter tables, code
+blocks, admonitions, download links, captions and navigation. For a docstring that
+is its summary and the paragraphs following it; for a gallery example it is the
+example's introduction. The plain ``description`` meta tag is set to match, unless
+``ogp_enable_meta_description`` is disabled.
